pyproject.toml
[build-system]
build-backend = "setuptools.build_meta"
requires = [ "setuptools" ]

[project]
name = "so-vits-svc-fork"
version = "4.2.29"
description = "A fork of so-vits-svc."
readme = "README.md"
license = { text = "MIT" }
authors = [
  { name = "34j", email = "34j.95a2p@simplelogin.com" },
]
requires-python = ">=3.9"
classifiers = [
  "Development Status :: 2 - Pre-Alpha",
  "Intended Audience :: Developers",
  "Natural Language :: English",
  "Operating System :: OS Independent",
  "Programming Language :: Python :: 3.9",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Programming Language :: Python :: 3.13",
  "Topic :: Software Development :: Libraries",
]

dependencies = [
    "click>=8.1.8",
    "cm-time>=0.1.2",
    "fastapi>=0.116.1",
    "librosa>=0.11.0",
    "lightning>=2.5.5",
    "matplotlib>=3.9.4",
    "numpy>=2.0.2",
    "pebble>=5.1.3",
    "praat-parselmouth>=0.4.6",
    "psutil>=7.1.2",
    "pywebview>=5.0.0",
    "pyworld>=0.3.5",
    "requests>=2.32.5",
    "rich>=14.1.0",
    "scipy>=1.13.1",
    "sounddevice>=0.5.2",
    "soundfile>=0.13.1",
    "tensorboard>=2.20.0",
    "tensorboardx>=2.6.4",
    "torch>=2.8.0",
    "torchaudio>=2.8.0",
    "torchcrepe>=0.0.24",
    "tqdm>=4.67.1",
    "tqdm-joblib>=0.0.4",
    "transformers>=4.56.1",
    "uvicorn>=0.34.0",
]
urls."Bug Tracker" = "https://github.com/voicepaw/so-vits-svc-fork/issues"
urls.Changelog = "https://github.com/voicepaw/so-vits-svc-fork/blob/main/CHANGELOG.md"
urls.documentation = "https://so-vits-svc-fork.readthedocs.io"
urls.repository = "https://github.com/voicepaw/so-vits-svc-fork"
scripts.svc = "so_vits_svc_fork.__main__:cli"
scripts.svcg = "so_vits_svc_fork.gui_web:main"

[dependency-groups]
dev = [
  "pytest>=8,<9",
  "pytest-cov>=7,<8",
]
docs = [
  "furo>=2023.5.20; python_version>='3.11'",
  "myst-parser>=0.16; python_version>='3.11'",
  "sphinx>=4; python_version>='3.11'",
  "sphinx-autobuild>=2025,<2026; python_version>='3.11'",
]

[tool.ruff]
line-length = 150
lint.select = [
  # "B",   # flake8-bugbear
  # "D",   # flake8-docstrings
  # "C4",  # flake8-comprehensions
  # "S",   # flake8-bandit
  "F",   # pyflake
  # "E",   # pycodestyle
  "W",   # pycodestyle
  # "UP",  # pyupgrade
  "I",   # isort
  # "RUF", # ruff specific
]
lint.ignore = [
  "D203", # 1 blank line required before class docstring
  "D212", # Multi-line docstring summary should start at the first line
  "D100", # Missing docstring in public module
  "D104", # Missing docstring in public package
  "D107", # Missing docstring in `__init__`
  "D401", # First line of docstring should be in imperative mood
]
lint.per-file-ignores."conftest.py" = [ "D100" ]
lint.per-file-ignores."docs/conf.py" = [ "D100" ]
lint.per-file-ignores."setup.py" = [ "D100" ]
lint.per-file-ignores."tests/**/*" = [
  "D100",
  "D101",
  "D102",
  "D103",
  "D104",
  "S101",
]
lint.isort.known-first-party = [ "so_vits_svc_fork", "tests" ]

[tool.pytest.ini_options]
addopts = """\
    -v
    -Wdefault
    --cov=so_vits_svc_fork
    --cov-report=term
    --cov-report=xml
    """
pythonpath = [ "src" ]

[tool.coverage.run]
branch = true

[tool.coverage.report]
exclude_lines = [
  "pragma: no cover",
  "@overload",
  "if TYPE_CHECKING",
  "raise NotImplementedError",
  'if __name__ == "__main__":',
]

[tool.mypy]
check_untyped_defs = true
disallow_any_generics = true
disallow_incomplete_defs = true
disallow_untyped_defs = true
mypy_path = "src/"
no_implicit_optional = true
show_error_codes = true
warn_unreachable = true
warn_unused_ignores = true
exclude = [
  'docs/.*',
  'setup.py',
]

[[tool.mypy.overrides]]
module = "tests.*"
allow_untyped_defs = true

[[tool.mypy.overrides]]
module = "docs.*"
ignore_errors = true

[tool.semantic_release]
version_toml = [ "pyproject.toml:project.version" ]
version_variables = [
  "src/so_vits_svc_fork/__init__.py:__version__",
  "docs/conf.py:release",
]
build_command = """
pip install uv
uv lock
git add uv.lock
uv build
"""

[tool.semantic_release.changelog]
exclude_commit_patterns = [
  '''chore(?:\([^)]*?\))?: .+''',
  '''ci(?:\([^)]*?\))?: .+''',
  '''refactor(?:\([^)]*?\))?: .+''',
  '''style(?:\([^)]*?\))?: .+''',
  '''test(?:\([^)]*?\))?: .+''',
  '''build\((?!deps\): .+)''',
  '''Merged? .*''',
  '''Initial [Cc]ommit.*''', # codespell:ignore
]

[tool.semantic_release.changelog.environment]
keep_trailing_newline = true

[tool.semantic_release.branches.main]
match = "main"

[tool.semantic_release.branches.noop]
match = "(?!main$)"
prerelease = true
